+++ sdk/signalr/Microsoft.Azure.WebJobs.Extensions.SignalRService/src/Client/IAzureSignalRSender.cs
@@ -1,12 +1,19 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
+using System.Collections.Generic;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
 {
     internal interface IAzureSignalRSender
     {
+        Task<SignalRConnectionInfo> RefreshConnectionInfoAsync(string connectionToken, DateTimeOffset expireTime, IList<Claim> claims);
+
+        Task<IList<Claim>> GetConnectionClaimsAsync(string connectionToken);
+
         Task SendToAll(SignalRData data);
 
         Task SendToConnection(string connectionId, SignalRData data);